complete components for server and credentials data manipulation, add UpsSelect and CredentialsSelect for dropdown selection

+++ frontend/src/components/ServerContentDisplayAndEditing.tsx
@@ -2,8 +2,10 @@ import {Ups} from "../types/ups.ts";
 import {useEffect, useState} from "react";
 import axios from "axios";
 import {useNavigate} from "react-router-dom";
-import { Server } from "../types/server.ts";
-import { Credentials } from "../types/credentials.ts";
+import {Server} from "../types/server.ts";
+import {Credentials} from "../types/credentials.ts";
+import CredentialsSelect from "./CredentialsSelect.tsx";
+import UpsSelect from "./UpsSelect.tsx";
 
 type EditProps = {
     server: Server,
@@ -15,30 +17,37 @@ type EditProps = {
 
 export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) {
 
-    const [server, setServer] = useState<Server>({id: "new", name: "", address: "", upsId: "", shutdownTime: 180, credentials:{id:"", user:"", password:"", global:false}})
+    const [server, setServer] = useState<Server>({
+        id: "new",
+        name: "",
+        address: "",
+        upsId: "",
+        shutdownTime: 180,
+        credentials: {id: "", user: "", password: "", global: false}
+    })
     const [editing, setEditing] = useState<boolean>(false)
     const [changedData, setChangedData] = useState<boolean>(false)
     const [nameInput, setNameInput] = useState<string>("")
     const [addressInput, setAddressInput] = useState<string>("")
-    const [userInput, setUserInput] = useState<string>("")
-    const [passwordInput, setPasswordInput] = useState<string>("")
     const [upsSelection, setUpsSelection] = useState<string>("")
     const [changedCredentialsInput, setChangedCredentialsInput] = useState<boolean>(false)
-    const [credentialsLocalSelected, setCredentialsLocalSelected] = useState<boolean>(false)
-    const [globalCredentialsSelection, setGlobalCredentialsSelection] = useState<Credentials>()
-    const [shutdownMinutesInput, setShutdownMinutesInput] = useState<number>(3)
+    const [localSelected, setLocalSelected] = useState<boolean>(true)
+    const [userInput, setUserInput] = useState<string>("")
+    const [passwordInput, setPasswordInput] = useState<string>("")
+    const [globalCredentialsSelection, setGlobalCredentialsSelection] = useState<string>("")
+    const [shutdownSecondsInput, setShutdownSecondsInput] = useState<number>(3)
     const [message, setMessage] = useState<string>("")          // in case of errors and for warning before deletion
     const confirmationMessage: string = "Really delete? Reclick button to confirm."
     const navigate = useNavigate()
 
-    function switchEditMode(state:boolean) : void {
+    function switchEditMode(state: boolean): void {
         setEditing(state)
         setChangedData(false)
         setChangedCredentialsInput(false)
         setMessage("")
     }
 
-    function backToList(updated:boolean) : void {
+    function backToList(updated: boolean): void {
         switchEditMode(false)
         if (updated) props.serverUpdate()
         navigate("/server")
@@ -50,74 +59,80 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
         setEditing(props.server.id === "new")
         setNameInput(props.server.name)
         setAddressInput(props.server.address)
-        if (props.server.credentials.global){
-            setCredentialsLocalSelected(false)
-            setGlobalCredentialsSelection(props.server.credentials)
+        if (props.server.credentials === null) {
+            setGlobalCredentialsSelection("")
+            setLocalSelected(true)
+        } else if (props.server.credentials.global) {
+            setLocalSelected(false)
+            setGlobalCredentialsSelection(props.server.credentials.id)
             setUserInput("")
             setPasswordInput("")
-        }
-        else {
-            setUserInput(props.server.credentials.user)
-            setPasswordInput(props.server.credentials.password)
-            setCredentialsLocalSelected(true)
-            setGlobalCredentialsSelection(undefined)
+        } else {
+            setUserInput(props.server.credentials?.user)
+            setPasswordInput(props.server.credentials?.password)
+            setLocalSelected(true)
+            setGlobalCredentialsSelection("")
         }
         setUpsSelection(props.server.upsId)
-        setShutdownMinutesInput(props.server.shutdownTime / 60)
+        setShutdownSecondsInput(props.server.shutdownTime)
     }, [props.server])
 
     /** reset form data to content of server object */
-    function resetForm(){
+    function resetForm() {
         setNameInput(server.name)
         setAddressInput(server.address)
-        if (server.credentials.global) {
-            setCredentialsLocalSelected(false)
-            setGlobalCredentialsSelection(props.server.credentials)
+        if (server.credentials?.global) {
+            setLocalSelected(false)
+            setGlobalCredentialsSelection(props.server.credentials.id)
             setUserInput("")
             setPasswordInput("")
-        }
-        else {
-            setCredentialsLocalSelected(true)
-            setUserInput(server.credentials.user)
-            setPasswordInput(server.credentials.password)
+        } else {
+            setLocalSelected(true)
+            setUserInput(server.credentials?.user)
+            setPasswordInput(server.credentials?.password)
         }
         setUpsSelection(server.upsId)
-        setShutdownMinutesInput(server.shutdownTime / 60)
+        setShutdownSecondsInput(server.shutdownTime)
         setChangedData(false)
         setChangedCredentialsInput(false)
     }
+
+    function getGlobalCredentialsById(id: string): Credentials | undefined {
+        return props.credentialsList.find(c => c.id === id)
+    }
+
 
     /** axios calls for credentials *********************************************************************************/
 
     /** @return new Credentials object from backend with provided data in case of success,
      *          null in case of failure
      */
-    function addLocalCredentials() : Credentials|null {
-        axios.post('api/credentials', {user: userInput, password: passwordInput, global: false})
+    function addLocalCredentials(): Credentials | undefined {
+        axios.post('/api/credentials', {user: userInput, password: passwordInput, global: false})
             .then(response => {
                 if (response.status == 200) {
                     return response.data
-                }
-                else {
+                } else {
                     setMessage(response.data)
-                    return null
                 }
             })
             .catch(error => {
                 console.error("addLocalredentials failed:", error)
-                return null
             })
-        return null
+        return undefined
     }
 
     /** @return true if update was successful, false otherwise */
-    function updateLocalCredentials() : boolean {
-        axios.post('api/credentials/' + server.credentials.id, {user: userInput, password: passwordInput, global: false})
+    function updateLocalCredentials(): boolean {
+        axios.post('/api/credentials/' + server.credentials.id, {
+            user: userInput,
+            password: passwordInput,
+            global: false
+        })
             .then(response => {
                 if (response.status == 200) {
                     return true
-                }
-                else {
+                } else {
                     setMessage(response.data)
                     return false
                 }
@@ -131,12 +146,13 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
 
     /** @return true if delete was successful, false otherwise */
     function deleteLocalCredentials(): boolean {
+        if (!server.credentials)  /** when credentials value was null, nothing needs to be deleted => success */
+            return true
         axios.delete('/api/credentials/' + server.credentials.id)
             .then(response => {
                 if (response.status == 200) {
                     return true
-                }
-                else {
+                } else {
                     setMessage(response.data)
                     return false
                 }
@@ -149,12 +165,13 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
     }
 
     /** preparation of embedded credentials object before server object can be stored **************************************/
-    function buildEmbeddedCredentialsObject(): Credentials|null {
-        if (credentialsLocalSelected) {
-            if (server.id === "new" || server.credentials.global) {
+    function buildEmbeddedCredentialsObject(): Credentials | undefined {
+        if (localSelected) {
+            if (server.id === "new" || !server.credentials || server.credentials.global) {
                 /** when for a new server local Credentials are chosen
-                 * or an existing server changes from global to local credentials
-                 *  a new Credentials object has to be created first
+                 *  (or for an existing server with no value for credentials)
+                 *  or an existing server changes from global to local credentials,
+                 *  then a new local Credentials object has to be created first
                  */
                 return addLocalCredentials()
             } else if (changedCredentialsInput) {
@@ -165,28 +182,26 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
                         id: server.credentials.id,
                         user: userInput,
                         password: passwordInput,
-                        global: false}
+                        global: false
+                    }
                     : server.credentials
-            }
-            else /** no changes to local credentials */
-                 return server.credentials
-        }
-        else
-            /** when global credentials are chosen, data is found in globalCredentialsSelection
-             * (if this should be undefined return null)
+            } else /** no changes to local credentials */
+                return server.credentials
+        } else
+            /** when global credentials are chosen, data is found in credentialsList
              */
-            return globalCredentialsSelection ? globalCredentialsSelection : null
+            return getGlobalCredentialsById(globalCredentialsSelection)
     }
 
     /** axios calls for server ********************************************************************/
-    function addServer():void{
-        const credentialsToStore : Credentials|null = buildEmbeddedCredentialsObject()
+    function addServer(): void {
+        const credentialsToStore: Credentials | undefined = buildEmbeddedCredentialsObject()
         axios.post('/api/server', {
             name: nameInput,
             address: addressInput,
             credentials: credentialsToStore,
             upsId: upsSelection,
-            shutdownTime: Math.round(shutdownMinutesInput * 60)
+            shutdownTime: shutdownSecondsInput
         })
             .then(response => {
                 if (response.status == 200)
@@ -198,23 +213,22 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
             });
     }
 
-    function updateServer():void {
+    function updateServer(): void {
         const credentialsToStore = buildEmbeddedCredentialsObject()
         axios.put('/api/server/' + server.id, {
             name: nameInput,
             address: addressInput,
             credentials: credentialsToStore,
             upsId: upsSelection,
-            shutdownTime: Math.round(shutdownMinutesInput * 60)
+            shutdownTime: shutdownSecondsInput
         })
             .then(response => {
-                if (response.status == 200){
+                if (response.status == 200) {
                     /** delete localCredentials if not contained in updated server object */
-                    if ((credentialsToStore===null || credentialsToStore.global) && !server.credentials.global)
+                    if ((!credentialsToStore || credentialsToStore.global) && !server.credentials?.global)
                         deleteLocalCredentials()
                     backToList(true)
-                }
-                else setMessage(response.data)
+                } else setMessage(response.data)
             })
             .catch(error => {
                 console.error('updateServer failed:', error);
@@ -229,13 +243,13 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
                     if (!server.credentials.global)
                         deleteLocalCredentials()
                     backToList(true)
-                }
-                else setMessage(response.data)
+                } else setMessage(response.data)
             })
             .catch(error => {
                 console.error('deleteServer failed:', error);
             });
     }
+
     /** end axios calls for server *************************************************************/
 
     function submitEditForm(): void {
@@ -243,18 +257,17 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
             setMessage("Error: Address is mandatory")
             return
         }
-        if (credentialsLocalSelected && (userInput==="" || passwordInput==="")){
+        if (localSelected && (userInput === "" || passwordInput === "")) {
             setMessage("For local users username and password must be entered")
             return
         }
-        if (shutdownMinutesInput < 0) {
+        if (shutdownSecondsInput < 0) {
             setMessage("Time for Shutdown cannot be negative")
             return;
         }
-        if (server.id==="new") {
+        if (server.id === "new") {
             addServer()
-        }
-        else {
+        } else {
             updateServer()
         }
     }
@@ -288,6 +301,20 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
         }}
     />;
 
+    const localCheckbox = <>
+        <input
+            id={"local"}
+            name={"local"}
+            type={"checkbox"}
+            checked={localSelected}
+            disabled={!editing}
+            onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
+                setLocalSelected(event.target.checked)
+            }}
+        />
+        <label htmlFor={"local"}> use local credentials (only on this server)</label>
+    </>;
+
     const userInputField = <input
         id={'user'}
         type={'text'}
@@ -312,13 +339,37 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
         }}
     />;
 
-    const shutdownMinutesField = <input
-        id={'minutes'}
+    const localUserInput = <>
+        <li>
+            <label htmlFor={'user'}>Username:</label>
+            {editing
+                ? userInputField
+                : <p>{server.credentials?.user}</p>}
+        </li>
+        <li>
+            <label htmlFor={'password'}>Password:</label>
+            {editing
+                ? passwordInputField
+                : <p>********</p>}
+        </li>
+    </>;
+
+    const globalUserInput = <>
+        <li>
+            <label>Global Credentials:</label>
+            <CredentialsSelect disabled={!editing} selection={globalCredentialsSelection}
+                               setSelected={setGlobalCredentialsSelection}
+                               setChangedData={setChangedData} credentialsList={props.credentialsList}/>
+        </li>
+    </>
+
+    const shutdownSecondsField = <input
+        id={'seconds'}
         type={'number'}
-        name={'minutes'}
-        value={shutdownMinutesInput}
+        name={'seconds'}
+        value={shutdownSecondsInput}
         onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-            setShutdownMinutesInput(parseFloat(event.target.value))
+            setShutdownSecondsInput(parseInt(event.target.value))
             setChangedData(true)
             setChangedCredentialsInput(true)
         }}
@@ -327,7 +378,7 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
     return (
         <>
             <h3>Details of Server</h3>
-            <button onClick={() => backToList(false)} >
+            <button onClick={() => backToList(false)}>
                 Back
             </button>
 
@@ -346,22 +397,24 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
                             : <p>{server.address}</p>}
                     </li>
                     <li>
-                        <label htmlFor={'user'}>Username:</label>
-                        {editing
-                            ? userInputField
-                            : <p>{server.credentials.user}</p>}
+                        {localCheckbox}
+                    </li>
+                    {localSelected
+                        ? localUserInput
+                        : globalUserInput
+                    }
+                    <li>
+                        <label>On UPS:</label>
+                        <UpsSelect disabled={!editing} selection={globalCredentialsSelection}
+                                           setSelected={setUpsSelection}
+                                           setChangedData={setChangedData} upsList={props.upses}/>
                     </li>
                     <li>
-                        <label htmlFor={'password'}>Password:</label>
+                        <label htmlFor={'seconds'}>Shutdown Trigger:</label>
                         {editing
-                            ? passwordInputField
-                            : <p>{server.credentials.user}</p>}
-                    </li>
-                    <li>
-                        <label htmlFor={'minutes'}>Remaining battery time for shutdown (minutes):</label>
-                        {editing
-                            ? shutdownMinutesField
-                            : <p>{server.credentials.user}</p>}
+                            ? shutdownSecondsField
+                            : <p>{server.shutdownTime}</p>}
+                        <label htmlFor={"seconds"}> seconds remaining battery time</label>
                     </li>
                     <li></li>
                     <li>
@@ -378,7 +431,8 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
                         <button type={"button"} onClick={deleteClicked} hidden={editing}>
                             Delete
                         </button>
-                        <button type={"button"} onClick={() => switchEditMode(true)} hidden={editing || message===confirmationMessage}>
+                        <button type={"button"} onClick={() => switchEditMode(true)}
+                                hidden={editing || message === confirmationMessage}>
                             Edit
                         </button>
                     </li>
@@ -388,5 +442,4 @@ export default function UpsContentDisplayAndEditing(props: Readonly<EditProps>) 
         </>
     )
 }
-
